Add files via upload
On my new 2023 Lenovo Chromebook, it didn’t take for the RATNinjas to comprise it. Using a (still) existing exploit, they modified the system to run crostini and effectively turned it into a mobile Android device (see images). Therefore, I am unable to run apps at full size (except Chrome the browser). This script contained information specific to my Google account(s). This hack is slightly adjusted to run on macOS and iOS, but this examples comes from a full-size Chromebook with keyboard. 

It appears by running in developer mode, they can effectively control all aspects of the laptop. The presence of .test files in this example suggests it is also collecting data from the session.
https://developer.android.com/kotlin/flow/test

CaptiveGoogle runs each of my families accounts (regardless of the device type) through a Google API. The splash page says Google Workspace but I suspect it’s not actually Workspace as they are subscription based and I only have a Google One subscription.

I have tried to reach out to Google, but without a fully reverse engineered example including the specific entry point they are not interested. The so saga continues for myself and many others at the mercy of our stealthy developer captors…